Added unit test for class WeatherService

+++ src/main/java/com/weather/app/model/WeatherHistory.java
@@ -19,6 +19,6 @@ public class WeatherHistory {
     private String postalCode;
     private String userName;
     private Double temperature;
-    private Double humidity;
+    private int humidity;
     private String weatherCondition;
 }
